refactor: 도서 csv data dto는 단순 String을 전달하고, 형변환은 processor에서 하도록 변경
