+++ src/main/java/com/bookspot/batch/book/processor/BookCsvProcessor.java
@@ -2,11 +2,15 @@ package com.bookspot.batch.book.processor;
 
 import com.bookspot.batch.book.data.Book;
 import com.bookspot.batch.book.reader.BookCsvData;
+import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class BookCsvProcessor implements ItemProcessor<BookCsvData, Book> {
+    private final YearParser yearParser;
+
     @Override
     public Book process(BookCsvData item) throws Exception {
         if(item.getIsbn13() == null)
@@ -16,7 +20,7 @@ public class BookCsvProcessor implements ItemProcessor<BookCsvData, Book> {
                 item.getIsbn13(),
                 item.getTitle(),
                 item.getAuthor(),
-                item.getPublicationYear(),
+                yearParser.parse(item.getPublicationYear()),
                 item.getSubjectCode()
         );
     }
